requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
asyncer==0.0.2
attrs==23.1.0
auth0-python==4.3.0
backoff==2.2.1
beautifulsoup4==4.12.2
bidict==0.22.1
certifi==2023.5.7
cffi==1.15.1
chainlit==0.5.1
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
courlan==0.9.3
cryptography==41.0.1
dataclasses-json==0.5.9
dateparser==1.1.8
Deprecated==1.2.14
exceptiongroup==1.1.2
fastapi==0.97.0
fastapi-socketio==0.0.10
filelock==3.12.2
filetype==1.2.0
frozenlist==1.3.3
fsspec==2023.6.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.56.0
h11==0.14.0
htmldate==1.4.3
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.0.1
Jinja2==3.1.2
jusText==3.0.0
langchain==0.0.229
langchainplus-sdk==0.0.20
langcodes==3.3.0
llama-index==0.7.4
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
nodeenv==1.8.0
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.18.0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-exporter-otlp-proto-common==1.18.0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-exporter-otlp-proto-http==1.18.0
opentelemetry-instrumentation==0.39b0
opentelemetry-proto==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
packaging==23.1
pandas==2.0.3
prisma==0.9.1
protobuf==4.23.4
pycparser==2.21
pydantic==1.10.11
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-graphql-client==0.4.3
python-socketio==5.8.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
starlette==0.27.0
syncer==2.0.3
tenacity==8.2.2
tiktoken==0.4.0
tld==0.13
tomli==2.0.1
tomlkit==0.11.8
tqdm==4.65.0
trafilatura==1.6.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
uptrace==1.18.0
urllib3==1.26.16
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2
zipp==3.16.0
